Fiexed Defect

+++ app/src/main/java/com/digimbanking/Features/Transfer/SesamaBank/InputNominalViewModel.kt
@@ -5,6 +5,7 @@ import com.core.data.repositories.AkunRepository
 import com.core.data.repositories.ProfilRepository
 import com.core.data.repositories.RiwayatRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
+import java.math.BigInteger
 import javax.inject.Inject
 
 @HiltViewModel
@@ -15,7 +16,7 @@ class InputNominalViewModel @Inject constructor(
 
     fun getUser() = apiService.getAkun()
     fun validateNoninal(nominal: String): Boolean{
-        isNominalValid = nominal.toLong() >= 10000L
+        isNominalValid = nominal.toBigInteger() >= BigInteger.valueOf(10000)
         return isNominalValid
     }
 }
